[Dani S, Dani C] refactoring of symptom registry list tests
